scripts/requirements.txt
boto3
requests
requests-aws4auth
python-dotenv
